"Under-the-hood" 0.7
Update Gradle configuration and ProGuard rules to disable minification and remove mobile-ffmpeg dependency

This commit modifies the build.gradle.kts file to disable resource shrinking and minification for the debug build type. It also updates the resources configuration to exclude specific META-INF files. Additionally, the ProGuard rules are adjusted to remove the retention of the mobile-ffmpeg library classes, and the corresponding dependency is removed from the libs.versions.toml file, streamlining the project dependencies.